+++ lib/objekts.dart
@@ -1,5 +1,6 @@
 export 'src/capture.dart' show screenshots;
 export 'src/font_loader.dart' show loadAppFonts;
+export 'src/golden.dart' show matchesGolden;
 export 'src/models.dart';
 export 'src/test_widgets.dart'
     show ObjektsDeviceWidgetTesterCallback, testWidgets, testWidgetsForDevices;